Move to commandStore & Playwright test
